+++ src/helpers/Html/Html.jsx
@@ -73,6 +73,8 @@ class Html extends Component {
 
           <link rel="shortcut icon" href="/favicon.ico" />
           <meta name="viewport" content="width=device-width, initial-scale=1" />
+          <link rel="manifest" href="/manifest.json" />
+          <meta name="theme-color" content="#000000" />
           <meta name="apple-mobile-web-app-capable" content="yes" />
           {assets.client.css ? (
             <link rel="stylesheet" href={assets.client.css} />